refactor(-): :hammer: refatora utils de dieta_especial
