+++ src/dieta_especial/utils/admin.py
@@ -1,0 +1,5 @@
+import re
+
+
+def is_alpha_numeric_and_has_single_space(descricao):
+    return bool(re.match(r"[A-Za-z0-9\s]+$", descricao))
